+++ src/file_converter_MD_to_HTML.py
@@ -11,9 +11,8 @@ class MDToHTMLConverter(FileConverter):
         self.rename_target_file_if_already_exists()
 
     def parse_metadata_if_required(self):
-        if not self._conversion_settings.markdown_conversion_input == 'pandoc_markdown':
-            self._metadata_processor = MetaDataProcessor(self._conversion_settings)
-            self._pre_processed_content = self._metadata_processor.parse_md_metadata(self._pre_processed_content)
+        self._metadata_processor = MetaDataProcessor(self._conversion_settings)
+        self._pre_processed_content = self._metadata_processor.parse_md_metadata(self._pre_processed_content)
 
     def post_process_content(self):
         self.logger.debug(f'Post process HTML content')
@@ -22,7 +21,4 @@ class MDToHTMLConverter(FileConverter):
         self.add_meta_data_if_required()
 
     def add_meta_data_if_required(self):
-        if self._conversion_settings.markdown_conversion_input == 'pandoc_markdown':
-            return
-
         self._post_processed_content = self._metadata_processor.add_metadata_html_to_content(self._post_processed_content)