+++ test/test_file_converter_HTML_to_MD.py
@@ -33,8 +33,9 @@ class TestHTMLToMDConverter(unittest.TestCase):
         self.file_converter._conversion_settings.export_format = 'pandoc_markdown'
         self.file_converter.pre_process_content()
         self.file_converter.convert_content()
+        self.file_converter._metadata_processor._conversion_settings.front_matter_format = 'toml'  # set toml and confirm content is forced back into yaml
         self.file_converter.post_process_content()
-        self.assertEqual('- [x] Check 1\n- [ ] Check 2\n<img src="filepath/image.png" width="600" />\n\n', self.file_converter._post_processed_content, 'post processing failed')
+        self.assertEqual('---\ntitle: this is test2\n---\n\n- [x] Check 1\n- [ ] Check 2\n<img src="filepath/image.png" width="600" />\n', self.file_converter._post_processed_content, 'post processing failed')
 
         self.file_converter._file_content = '<head><meta title="this is test2"/><meta not_valid="not_in_schema"/></head><p><input checked="" type="checkbox"/>Check 1</p><p><input type="checkbox"/>Check 2</p><img src="filepath/image.png" width="600">'
         self.file_converter._metadata_schema = ['title']
@@ -71,4 +72,4 @@ class TestHTMLToMDConverter(unittest.TestCase):
         self.file_converter._metadata_processor._metadata_schema = ['title', 'creation_time']
         self.file_converter._pre_processed_content = '<head><meta title="this is test2"/><meta creation_time="test-meta-content"/></head>'
         self.file_converter.parse_metadata_if_required()
-        self.assertEqual({}, self.file_converter._metadata_processor.metadata, 'meta data not parsed correctly')
+        self.assertEqual({'title': 'this is test2'}, self.file_converter._metadata_processor.metadata, 'meta data not parsed correctly')